net/raft: move from database/raft
