+++ net/raft/rand.go
